@@ -25,12 +25,12 @@
         "dva-loading": "^2.0.3",
         "history": "^4.7.2",
         "install": "^0.12.1",
-        "lodash": "^4.17.10",
+        "lodash": "^4.17.15",
         "noform": "^1.1.3",
         "prop-types": "^15.6.2",
         "rc-queue-anim": "^1.6.8",
-        "react": "^16.2.0",
-        "react-dom": "^16.2.0"
+        "react": "^16.9.0",
+        "react-dom": "^16.9.0"
     },
     "devDependencies": {
         "antd": "^3.6.6",
